+++ Domain/Abstractions/IHotelRepository.cs
@@ -14,8 +14,8 @@ namespace Domain.Abstractions
             (
                 int page,
                 int pageSize,
-                int minPrice,
-                int? maxPrice,
+                decimal minPrice,
+                decimal? maxPrice,
                 string? city,
                 string? country,
                 HotelType[] hotelType,
